+++ sim/lib/settings/town_settings_heinsberg.py
@@ -21,13 +21,13 @@ population_per_age_group = np.array([
     55436,  # 15-34
     91858,  # 35-59
     55703,  # 60-79
-    15555]) # 80+
+    15555], dtype=np.int32)  # 80+
 
 town_population = 41673 # Population of the central town of the region
 region_population = population_per_age_group.sum()
 
 # Roughly 100k in total in Germany: https://www.rki.de/DE/Content/Infekt/EpidBull/Archiv/2020/Ausgaben/15_20.pdf?__blob=publicationFile
-daily_tests_unscaled = int(100000 * town_population / 83000000)
+daily_tests_unscaled = int(100000 * (town_population / 83000000))
 
 # Information about household structure (set to None if not available)
 # Source for Germany: https://www.destatis.de/EN/Themes/Society-Environment/Population/Households-Families/Tables/lrbev05.html
